+++ packages/insomnia/src/ui/components/tags/method-tag.tsx
@@ -1,11 +1,25 @@
 import React, { FC, memo } from 'react';
 
-import * as util from '../../../common/misc';
+import { METHOD_DELETE, METHOD_OPTIONS } from '../../../common/constants';
 
 interface Props {
   method: string;
   override?: string | null;
   fullNames?: boolean;
+}
+function removeVowels(str: string) {
+  return str.replace(/[aeiouyAEIOUY]/g, '');
+}
+function formatMethodName(method: string) {
+  let methodName = method || '';
+
+  if (method === METHOD_DELETE || method === METHOD_OPTIONS) {
+    methodName = method.slice(0, 3);
+  } else if (method.length > 4) {
+    methodName = removeVowels(method).slice(0, 4);
+  }
+
+  return methodName;
 }
 
 export const MethodTag: FC<Props> = memo(({ method, override, fullNames }) => {
@@ -13,8 +27,8 @@ export const MethodTag: FC<Props> = memo(({ method, override, fullNames }) => {
   let overrideName = override;
 
   if (!fullNames) {
-    methodName = util.formatMethodName(method);
-    overrideName = override ? util.formatMethodName(override) : override;
+    methodName = formatMethodName(method);
+    overrideName = override ? formatMethodName(override) : override;
   }
 
   return (
